Fix circular import in login
